+++ core/src/main/java/org/core/common/constant/PicTypeConstant.java
@@ -5,27 +5,27 @@ public class PicTypeConstant {
     /**
      * 歌曲
      */
-    public static final Short MUSIC = 0;
+    public static final Byte MUSIC = 0;
     
     /**
      * 歌单
      */
-    public static final Short COLLECT = 1;
+    public static final Byte COLLECT = 1;
     
     /**
      * 专辑
      */
-    public static final Short ALBUM = 2;
+    public static final Byte ALBUM = 2;
     
     /**
      * 歌手
      */
-    public static final Short ARTIST = 3;
+    public static final Byte ARTIST = 3;
     
     /**
      * 用户
      */
-    public static final Short USER = 4;
+    public static final Byte USER = 4;
     
     private PicTypeConstant() {
     }
