+++ core/src/main/java/org/core/common/constant/TargetTagConstant.java
@@ -4,22 +4,22 @@ public class TargetTagConstant {
     /**
      * 流派
      */
-    public static final Short TARGET_GENRE = 0;
+    public static final Byte TARGET_GENRE = 0;
     
     /**
      * 歌曲tag
      */
-    public static final Short TARGET_MUSIC_TAG = 1;
+    public static final Byte TARGET_MUSIC_TAG = 1;
     
     /**
      * 专辑流派
      */
-    public static final Short TARGET_ALBUM_GENRE = 0;
+    public static final Byte TARGET_ALBUM_GENRE = 0;
     
     /**
      * 歌单tag
      */
-    public static final Short TARGET_COLLECT_TAG = 2;
+    public static final Byte TARGET_COLLECT_TAG = 2;
     
     private TargetTagConstant() {
     }
