+++ core/src/main/java/org/core/config/PlayListTypeConfig.java
@@ -5,17 +5,17 @@ public class PlayListTypeConfig {
     /**
      * 普通歌单
      */
-    public static final Short ORDINARY = 0;
+    public static final byte ORDINARY = 0;
     
     /**
      * 喜爱歌单
      */
-    public static final Short LIKE = 1;
+    public static final byte LIKE = 1;
     
     /**
      * 推荐歌单
      */
-    public static final Short RECOMMEND = 2;
+    public static final byte RECOMMEND = 2;
     
     private PlayListTypeConfig() {
     }
