+++ plugin/src/main/java/org/plugin/common/TaskStatus.java
@@ -2,11 +2,11 @@ package org.plugin.common;
 
 public class TaskStatus {
     // 运行停止
-    public static final short STOP_STATUS = (short) 0;
+    public static final byte STOP_STATUS = 0;
     // 运行中
-    public static final short RUN_STATUS = (short) 1;
+    public static final byte RUN_STATUS = 1;
     // 运行错误
-    public static final short ERROR_STATUS = (short) 2;
+    public static final byte ERROR_STATUS = 2;
     
     private TaskStatus() {
     }
